+++ src/components/index.tsx
@@ -1,0 +1,12 @@
+import UserList, { UserListProps } from './common/UserList';
+import GlobisComponent from './Globis';
+import Home from './Home';
+import UserSearchForm from './UserSearchForm';
+
+export {
+  UserList,
+  UserListProps,
+  GlobisComponent,
+  Home,
+  UserSearchForm,
+};
